+++ main.py
@@ -1,7 +1,7 @@
 import pvleopard as pv, tqdm as t, os, shutil as s, core.scripts.blacklist as blacklist, core.scripts.fileScan as fs, core.scripts.spellCheck as sc, random as r
 from simple_image_download import simple_image_download as simp
 
-vnum = "1.6.5"
+vnum = "1.7"
 
 x = open("core/assets/logo.txt", "r")
 print(x.read())
@@ -72,10 +72,13 @@ try:
       try:
         for i in t.tqdm(range(0, 1), desc = "[INFO] downloading '" + query + "' " + str(ie) + "/" + str(len(all))):
           response().download(query, 10)
-        fname = query + "_" + str(r.randint(5, 10)) +".jpg"
-        s.move("simple_images/" + fname, "output/" + fname)
-        os.rename("output/" + fname, "output/" + query + ".jpg")
-        fs.checkFiles("output")
+        fs.checkFiles("simple_images/" + query)
+        working = os.listdir("simple_images/" + query)
+        for i in range(1,4):
+          working.remove(query + "_" + str(i) + ".jpg")
+        finalFile = str(r.choice(working))
+        s.move("simple_images/" + query + "/" + finalFile, "output/" + finalFile)
+        os.rename("output/" + finalFile, "output/" + query + ".jpg")
         ie = ie + 1
       except Exception as e:
         print("[INFO] " + str(e) + " " + str(ie) + "/" + str(len(all)))