+++ changelog.md
@@ -1,4 +1,5 @@
-# v1.6
-- Added spell check
-- Added extra error info
+# v1.7
+- Added image randomizer
+- Added file filter
 - Reorginized files
+- Enhanced fileScan.py